+++ raiden/tests/unit/api/test_python_api.py
@@ -3,7 +3,6 @@ import pytest
 
 from raiden.tests.utils.transfer import (
     channel,
-    direct_transfer,
 )
 from raiden.exceptions import (
     NoPathError,
@@ -43,75 +42,6 @@ def test_get_channel_list(raiden_network, token_addresses):
         )
 
 
-@pytest.mark.parametrize('channels_per_node', [0])
-@pytest.mark.parametrize('number_of_nodes', [2])
-@pytest.mark.parametrize('register_tokens', [False])
-@pytest.mark.parametrize('number_of_tokens', [1])
-def test_register_token(raiden_chain, token_addresses):
-    app0, _ = raiden_chain  # pylint: disable=unbalanced-tuple-unpacking
-
-    api0 = RaidenAPI(app0.raiden)
-    assert api0.manager_address_if_token_registered(token_addresses[0]) is None
-
-    manager_0token = api0.register_token(token_addresses[0])
-
-    assert manager_0token == api0.manager_address_if_token_registered(token_addresses[0])
-
-    # Exception if we try to reregister
-    with pytest.raises(ValueError):
-        api0.register_token(token_addresses[0])
-
-
-@pytest.mark.parametrize('channels_per_node', [0])
-@pytest.mark.parametrize('number_of_nodes', [2])
-@pytest.mark.parametrize('register_tokens', [False])
-@pytest.mark.parametrize('number_of_tokens', [1])
-def test_second_manager_address_if_token_registered(raiden_chain, token_addresses):
-    """Test recreating the scenario described on issue:
-    https://github.com/raiden-network/raiden/issues/784"""
-    app0, app1 = raiden_chain  # pylint: disable=unbalanced-tuple-unpacking
-
-    api0 = RaidenAPI(app0.raiden)
-    api1 = RaidenAPI(app1.raiden)
-
-    # Recreate the race condition by making sure the non-registering app won't
-    # register at all by watching for the TokenAdded blockchain event.
-    app1.raiden.alarm.remove_callback(app1.raiden.poll_blockchain_events)
-
-    manager_0token = api0.register_token(token_addresses[0])
-    # The second node does not register but just confirms token is registered.
-    # This is the behaviour the api call implement in register_token().
-    manager_1token = api1.manager_address_if_token_registered(token_addresses[0])
-
-    assert manager_0token == manager_1token
-
-    # Now make sure the token lists are populated for both nodes
-    tokens0_list = api0.get_tokens_list()
-    tokens1_list = api1.get_tokens_list()
-    assert tokens0_list == tokens1_list
-    assert len(tokens1_list) == 1
-    assert token_addresses[0] == tokens1_list[0]
-
-
-@pytest.mark.parametrize('channels_per_node', [1])
-@pytest.mark.parametrize('number_of_nodes', [2])
-@pytest.mark.parametrize('number_of_tokens', [1])
-def test_deposit_updates_balance_immediately(raiden_chain, token_addresses):
-    """Test that the balance of a channel gets updated by the deposit() call
-    immediately and without having to wait for the `ContractReceiveBalance`
-    message since the API needs to return the channel with the deposit balance
-    updated"""
-    app0, app1 = raiden_chain  # pylint: disable=unbalanced-tuple-unpacking
-
-    api0 = RaidenAPI(app0.raiden)
-
-    token_address = token_addresses[0]
-    channel_0_1 = channel(app0, app1, token_address)
-    old_balance = channel_0_1.contract_balance
-    returned_channel = api0.deposit(token_address, app1.raiden.address, 10)
-    assert returned_channel.contract_balance == old_balance + 10
-
-
 @pytest.mark.parametrize('blockchain_type', ['tester'])
 @pytest.mark.parametrize('number_of_nodes', [2])
 @pytest.mark.parametrize('channels_per_node', [1])
@@ -132,45 +62,6 @@ def test_transfer_to_unknownchannel(raiden_network):
             target='\xf0\xef3\x01\xcd\xcfe\x0f4\x9c\xf6d\xa2\x01?X4\x84\xa9\xf1',
             timeout=10
         )
-
-
-
-
-@pytest.mark.parametrize('blockchain_type', ['geth'])
-@pytest.mark.parametrize('channels_per_node', [1])
-@pytest.mark.parametrize('number_of_nodes', [2])
-def test_api_channel_events(raiden_chain):
-    app0, app1 = raiden_chain  # pylint: disable=unbalanced-tuple-unpacking
-    token_address = app0.raiden.chain.default_registry.token_addresses()[0]
-    channel_0_1 = channel(app0, app1, token_address)
-
-    amount = 30
-    direct_transfer(
-        app0,
-        app1,
-        token_address,
-        amount
-    )
-
-    results = RaidenAPI(app0.raiden).get_channel_events(channel_0_1.channel_address, 0)
-    assert len(results) == 3
-    max_block = 0
-    for idx, result in enumerate(results):
-        if result['block_number'] > max_block:
-            max_block = result['block_number']
-            assert max_block != 0
-
-        if idx == 2:
-            assert result['_event_type'] == 'EventTransferSentSuccess'
-        else:
-            assert result['_event_type'] == 'ChannelNewBalance'
-
-    assert max_block != 0
-
-    results = RaidenAPI(app0.raiden).get_channel_events(
-        channel_0_1.channel_address, max_block + 1, max_block + 100
-    )
-    assert not results
 
 
 @pytest.mark.parametrize('blockchain_type', ['tester'])
